+++ src/stpInterp/stpBetterTS.hpp
@@ -2,8 +2,9 @@
 #include "tree_sitter/api.h"
 
 #include <string>
+#include <cstdint>
 
 inline TSNode ts_node_child_by_field_name(const TSNode& self, const std::string& fieldName)
 {
-    return ts_node_child_by_field_name(self, fieldName.c_str(), fieldName.length());
+    return ts_node_child_by_field_name(self, fieldName.c_str(), static_cast<uint32_t>(fieldName.length()));
 }
